refactor: merge redundant default-images and placeholder-images libs
